requirements.txt
asgiref==3.4.1
coverage==6.0.1
Django==3.2.5
django-bootstrap-pagination==1.7.1
django-filter==21.1
django-widget-tweaks==1.4.9
Faker==9.8.0
humanize==4.0.0
libgravatar==1.0.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
typing-extensions==3.10.0.2
coverage==6.0.1
Django==3.2.5
django-bootstrap-pagination==1.7.1
django-filter==21.1
django-widget-tweaks==1.4.9
Faker==9.8.0
humanize==4.0.0
libgravatar==1.0.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
typing-extensions==3.10.0.2
pandas
django-with-asserts==0.0.1
scikit-surprise==1.1.1
scipy==1.8.0
joblib==1.1.0
gunicorn
django-heroku
